shim: Fix supported_cpuid copy from user
We need to copy_from_user in two operations. First to get the nent
field value, the number of entries. Then a second time with the
calculated size based on nent.

Copying back to user needs to include the padding field as well
otherwise copy to user is 4 bytes short.

Signed-off-by: Christopher Pelloux <git@chp.io>