vmm: mv_vs_op_msr_get_list handles unsupported MSR
Instead of failing when encountering an MSR that is not supported, we
now set the upper one bit of the 64 bit mv_rdl_entry_t.reg. MSR
registers are 32 bits index allowing us to reuse the upper 32 bits
without issues.

This is a temporary fix until we get support for the requested MSRs from
Qemu.