shim: Fix copying size in kvm_get_msr_index_list
On the first copy_from_user, we don't yet know what the total size of
the user provided kvm_msr_list object is, so we cannot copy_from_user
using our object size which might be different, instead we need to only
copy the size of the first field, nmsrs, allowing us to calculate the
size.

Due to AUTOSARS requirements, we are not allowed to use zero-length
arrays, effectively forcing us to not use the KVM API as is.
Our current approach seems to be error-prone as judging by the
number of recent fixes touching these structs and sizes given to
copy_{from,to}_user. A better approach might be desired.